refactor: colocate family achievement progress service tests
