+++ lib/family-achievement-progress-service.unlock.test.ts
@@ -1,7 +1,6 @@
-import { FamilyAchievementProgressService } from "../family-achievement-progress-service";
+import { FamilyAchievementProgressService } from "./family-achievement-progress-service";
 import type { MockChain } from "./family-achievement-progress-service.fixtures";
 import {
-  makeCountResult,
   makeDataResult,
   makeUpsertResult,
   makeUpdateResult,
@@ -18,12 +17,12 @@ jest.mock("@/lib/supabase-server", () => ({
   createServiceSupabaseClient: jest.fn(() => mockWriteClient),
 }));
 
-describe("FamilyAchievementProgressService — backfillProgress", () => {
+describe("FamilyAchievementProgressService — unlock evaluation", () => {
   afterEach(() => {
     jest.clearAllMocks();
   });
 
-  it("creates progress rows for all achievements regardless of event type", async () => {
+  it("sets unlocked_at when current >= threshold and not yet unlocked", async () => {
     const profilesChain = makeDataResult([
       { id: "user-001", characters: [{ id: "char-001" }] },
     ]);
@@ -39,18 +38,19 @@ describe("FamilyAchievementProgressService — backfillProgress", () => {
       },
     ]);
 
-    // No existing progress rows — simulates a family with no progress yet
-    const emptyProgressChain = makeDataResult([]);
+    const famProgressChain = makeDataResult([
+      { family_achievement_id: FAMILY_ACHIEVEMENT_ID },
+    ]);
 
     const questChain: MockChain = {
       select: jest.fn().mockReturnValue({
         eq: jest.fn().mockReturnValue({
-          or: jest.fn().mockResolvedValue({ count: 3, error: null }),
+          or: jest.fn().mockResolvedValue({ count: 5, error: null }),
         }),
       }),
     };
 
-    // Unlock check after upsert — achievement not yet unlocked
+    // Unlock check: not yet unlocked
     const unlockCheckChain: MockChain = {
       select: jest.fn().mockReturnValue({
         eq: jest.fn().mockReturnValue({
@@ -70,19 +70,21 @@ describe("FamilyAchievementProgressService — backfillProgress", () => {
     const readClient = makeReadClient({
       user_profiles: profilesChain as unknown as MockChain,
       family_achievements: famAchChain as unknown as MockChain,
+      family_achievement_progress: famProgressChain as unknown as MockChain,
       quest_instances: questChain,
     });
 
     let readProgressCallCount = 0;
-    const originalFrom = readClient.from as jest.Mock;
+    const originalFrom = readClient.from;
     readClient.from = jest.fn((table: string) => {
       if (table === "family_achievement_progress") {
         readProgressCallCount++;
-        return readProgressCallCount <= 1
-          ? emptyProgressChain
-          : unlockCheckChain;
+        if (readProgressCallCount <= 1) {
+          return (originalFrom as jest.Mock)(table);
+        }
+        return unlockCheckChain;
       }
-      return originalFrom(table);
+      return (originalFrom as jest.Mock)(table);
     }) as jest.Mock;
 
     const writeUpsert = makeUpsertResult();
@@ -90,106 +92,22 @@ describe("FamilyAchievementProgressService — backfillProgress", () => {
     let writeCallCount = 0;
     mockWriteClient.from.mockImplementation(() => {
       writeCallCount++;
-      return writeCallCount === 1 ? writeUpsert : writeUpdate;
+      if (writeCallCount === 1) return writeUpsert;
+      return writeUpdate;
     });
 
     const service = new FamilyAchievementProgressService(readClient as never);
-    await service.backfillProgress(FAMILY_ID);
+    await service.updateProgress(FAMILY_ID, { type: "QUEST_APPROVED" });
 
-    // Should have upserted a progress row
-    expect(writeUpsert.upsert).toHaveBeenCalledWith(
-      expect.arrayContaining([
-        expect.objectContaining({
-          family_id: FAMILY_ID,
-          family_achievement_id: FAMILY_ACHIEVEMENT_ID,
-        }),
-      ]),
-      expect.any(Object),
+    // Should have called upsert for progress
+    expect(writeUpsert.upsert).toHaveBeenCalled();
+    // Should have called update for unlock
+    expect(writeUpdate.update).toHaveBeenCalledWith(
+      expect.objectContaining({ unlocked_at: expect.any(String) }),
     );
   });
 
-  it("evaluates all criteria even when progress rows already exist", async () => {
-    // When backfillProgress is called and rows already exist, needsBackfill=false
-    // but event=null must still force all-criteria evaluation
-    const profilesChain = makeDataResult([
-      { id: "user-001", characters: [{ id: "char-001" }] },
-    ]);
-
-    const famAchChain = makeDataResult([
-      {
-        id: FAMILY_ACHIEVEMENT_ID,
-        name: "Test",
-        criteria_type: "reward_redeemed",
-        criteria_config: { threshold: 2, family_evaluation_mode: "sum" },
-        xp_reward: 0,
-        gold_reward: 0,
-      },
-    ]);
-
-    // Row already exists — needsBackfill would be false without explicit null event
-    const existingProgressChain = makeDataResult([
-      { family_achievement_id: FAMILY_ACHIEVEMENT_ID },
-    ]);
-
-    const rewardChain = makeCountResult(4);
-
-    const unlockCheckChain: MockChain = {
-      select: jest.fn().mockReturnValue({
-        eq: jest.fn().mockReturnValue({
-          in: jest.fn().mockResolvedValue({
-            data: [
-              {
-                family_achievement_id: FAMILY_ACHIEVEMENT_ID,
-                unlocked_at: null,
-              },
-            ],
-            error: null,
-          }),
-        }),
-      }),
-    };
-
-    const readClient = makeReadClient({
-      user_profiles: profilesChain as unknown as MockChain,
-      family_achievements: famAchChain as unknown as MockChain,
-      reward_redemptions: rewardChain,
-    });
-
-    let readProgressCallCount = 0;
-    const originalFrom = readClient.from as jest.Mock;
-    readClient.from = jest.fn((table: string) => {
-      if (table === "family_achievement_progress") {
-        readProgressCallCount++;
-        return readProgressCallCount <= 1
-          ? existingProgressChain
-          : unlockCheckChain;
-      }
-      return originalFrom(table);
-    }) as jest.Mock;
-
-    const writeUpsert = makeUpsertResult();
-    const writeUpdate = makeUpdateResult();
-    let writeCallCount = 0;
-    mockWriteClient.from.mockImplementation(() => {
-      writeCallCount++;
-      return writeCallCount === 1 ? writeUpsert : writeUpdate;
-    });
-
-    const service = new FamilyAchievementProgressService(readClient as never);
-    await service.backfillProgress(FAMILY_ID);
-
-    // reward_redeemed is not in QUEST_APPROVED criteria — only evaluated when event=null
-    expect(writeUpsert.upsert).toHaveBeenCalledWith(
-      expect.arrayContaining([
-        expect.objectContaining({
-          family_achievement_id: FAMILY_ACHIEVEMENT_ID,
-        }),
-      ]),
-      expect.any(Object),
-    );
-  });
-
-  it("sets unlocked_at when backfill finds threshold already met", async () => {
+  it("does NOT re-unlock already unlocked achievements", async () => {
     const profilesChain = makeDataResult([
       { id: "user-001", characters: [{ id: "char-001" }] },
     ]);
@@ -199,22 +117,25 @@ describe("FamilyAchievementProgressService — backfillProgress", () => {
         id: FAMILY_ACHIEVEMENT_ID,
         name: "Test",
         criteria_type: "quest_complete",
-        criteria_config: { threshold: 3, family_evaluation_mode: "sum" },
+        criteria_config: { threshold: 5, family_evaluation_mode: "sum" },
         xp_reward: 0,
         gold_reward: 0,
       },
     ]);
 
-    const emptyProgressChain = makeDataResult([]);
+    const famProgressChain = makeDataResult([
+      { family_achievement_id: FAMILY_ACHIEVEMENT_ID },
+    ]);
 
     const questChain: MockChain = {
       select: jest.fn().mockReturnValue({
         eq: jest.fn().mockReturnValue({
-          or: jest.fn().mockResolvedValue({ count: 5, error: null }), // current >= threshold
+          or: jest.fn().mockResolvedValue({ count: 10, error: null }),
         }),
       }),
     };
 
+    // Already unlocked
     const unlockCheckChain: MockChain = {
       select: jest.fn().mockReturnValue({
         eq: jest.fn().mockReturnValue({
@@ -222,7 +143,7 @@ describe("FamilyAchievementProgressService — backfillProgress", () => {
             data: [
               {
                 family_achievement_id: FAMILY_ACHIEVEMENT_ID,
-                unlocked_at: null,
+                unlocked_at: "2026-01-01T00:00:00Z",
               },
             ],
             error: null,
@@ -234,20 +155,69 @@ describe("FamilyAchievementProgressService — backfillProgress", () => {
     const readClient = makeReadClient({
       user_profiles: profilesChain as unknown as MockChain,
       family_achievements: famAchChain as unknown as MockChain,
+      family_achievement_progress: famProgressChain as unknown as MockChain,
       quest_instances: questChain,
     });
 
     let readProgressCallCount = 0;
-    const originalFrom = readClient.from as jest.Mock;
+    const originalFrom = readClient.from;
     readClient.from = jest.fn((table: string) => {
       if (table === "family_achievement_progress") {
         readProgressCallCount++;
-        return readProgressCallCount <= 1
-          ? emptyProgressChain
-          : unlockCheckChain;
+        if (readProgressCallCount <= 1) {
+          return (originalFrom as jest.Mock)(table);
+        }
+        return unlockCheckChain;
       }
-      return originalFrom(table);
+      return (originalFrom as jest.Mock)(table);
     }) as jest.Mock;
+
+    const writeUpsert = makeUpsertResult();
+    mockWriteClient.from.mockReturnValue(writeUpsert);
+
+    const service = new FamilyAchievementProgressService(readClient as never);
+    await service.updateProgress(FAMILY_ID, { type: "QUEST_APPROVED" });
+
+    // Only upsert should be called, not update (no unlock)
+    expect(writeUpsert.upsert).toHaveBeenCalled();
+    // mockWriteClient.from should only be called once (for upsert)
+    expect(mockWriteClient.from).toHaveBeenCalledTimes(1);
+  });
+
+  it("does NOT unlock when current < threshold", async () => {
+    const profilesChain = makeDataResult([
+      { id: "user-001", characters: [{ id: "char-001" }] },
+    ]);
+
+    const famAchChain = makeDataResult([
+      {
+        id: FAMILY_ACHIEVEMENT_ID,
+        name: "Test",
+        criteria_type: "quest_complete",
+        criteria_config: { threshold: 50, family_evaluation_mode: "sum" },
+        xp_reward: 0,
+        gold_reward: 0,
+      },
+    ]);
+
+    const famProgressChain = makeDataResult([
+      { family_achievement_id: FAMILY_ACHIEVEMENT_ID },
+    ]);
+
+    const questChain: MockChain = {
+      select: jest.fn().mockReturnValue({
+        eq: jest.fn().mockReturnValue({
+          or: jest.fn().mockResolvedValue({ count: 3, error: null }),
+        }),
+      }),
+    };
+
+    const readClient = makeReadClient({
+      user_profiles: profilesChain as unknown as MockChain,
+      family_achievements: famAchChain as unknown as MockChain,
+      family_achievement_progress: famProgressChain as unknown as MockChain,
+      quest_instances: questChain,
+    });
 
     const writeUpsert = makeUpsertResult();
     const writeUpdate = makeUpdateResult();
@@ -258,10 +228,13 @@ describe("FamilyAchievementProgressService — backfillProgress", () => {
     });
 
     const service = new FamilyAchievementProgressService(readClient as never);
-    await service.backfillProgress(FAMILY_ID);
+    await service.updateProgress(FAMILY_ID, { type: "QUEST_APPROVED" });
 
-    // Should have called update to set unlocked_at
-    expect(writeUpdate.update).toHaveBeenCalledWith(
+    // Upsert called, re-lock update called (to clear unlocked_at),
+    // but NOT the unlock update (which would set unlocked_at to a timestamp)
+    expect(writeUpsert.upsert).toHaveBeenCalled();
+    expect(writeUpdate.update).toHaveBeenCalledWith({ unlocked_at: null });
+    expect(writeUpdate.update).not.toHaveBeenCalledWith(
       expect.objectContaining({ unlocked_at: expect.any(String) }),
     );
   });
